+++ cleverhans/attacks/__init__.py
@@ -12,6 +12,8 @@ from cleverhans.attacks.fast_gradient_method import FastGradientMethod, fgm, opt
 from cleverhans.attacks.madry_et_al import MadryEtAl
 from cleverhans.attacks.momentum_iterative_method import MomentumIterativeMethod
 from cleverhans.attacks.projected_gradient_descent import ProjectedGradientDescent
+from cleverhans.attacks.saliency_map_method import SaliencyMapMethod
+from cleverhans.attacks.virtual_adversarial_method import VirtualAdversarialMethod, vatm
 from cleverhans.attacks_tf import SPSAAdam, margin_logit_loss, TensorAdam
 from cleverhans.model import Model, CallableModelWrapper
 from cleverhans.model import wrapper_warning, wrapper_warning_logits
@@ -23,252 +25,6 @@ from cleverhans import utils_tf
 
 _logger = utils.create_logger("cleverhans.attacks")
 tf_dtype = tf.as_dtype('float32')
-
-class SaliencyMapMethod(Attack):
-  """
-  The Jacobian-based Saliency Map Method (Papernot et al. 2016).
-  Paper link: https://arxiv.org/pdf/1511.07528.pdf
-
-  :param model: cleverhans.model.Model
-  :param sess: optional tf.Session
-  :param dtypestr: dtype of the data
-  :param kwargs: passed through to super constructor
-
-  :note: When not using symbolic implementation in `generate`, `sess` should
-         be provided
-  """
-
-  def __init__(self, model, sess=None, dtypestr='float32', **kwargs):
-    """
-    Create a SaliencyMapMethod instance.
-    Note: the model parameter should be an instance of the
-    cleverhans.model.Model abstraction provided by CleverHans.
-    """
-
-    super(SaliencyMapMethod, self).__init__(model, sess, dtypestr, **kwargs)
-
-    self.feedable_kwargs = ('y_target',)
-    self.structural_kwargs = [
-        'theta', 'gamma', 'clip_max', 'clip_min', 'symbolic_impl'
-    ]
-
-  def generate(self, x, **kwargs):
-    """
-    Generate symbolic graph for adversarial examples and return.
-
-    :param x: The model's symbolic inputs.
-    :param kwargs: See `parse_params`
-    """
-    # Parse and save attack-specific parameters
-    assert self.parse_params(**kwargs)
-
-    if self.symbolic_impl:
-      from cleverhans.attacks_tf import jsma_symbolic
-
-      # Create random targets if y_target not provided
-      if self.y_target is None:
-        from random import randint
-
-        def random_targets(gt):
-          result = gt.copy()
-          nb_s = gt.shape[0]
-          nb_classes = gt.shape[1]
-
-          for i in xrange(nb_s):
-            result[i, :] = np.roll(result[i, :],
-                                   randint(1, nb_classes - 1))
-
-          return result
-
-        labels, nb_classes = self.get_or_guess_labels(x, kwargs)
-        self.y_target = tf.py_func(random_targets, [labels],
-                                   self.tf_dtype)
-        self.y_target.set_shape([None, nb_classes])
-
-      x_adv = jsma_symbolic(
-          x,
-          model=self.model,
-          y_target=self.y_target,
-          theta=self.theta,
-          gamma=self.gamma,
-          clip_min=self.clip_min,
-          clip_max=self.clip_max)
-    else:
-      assert self.sess is not None, \
-        'Cannot use `generate` when no `sess` was provided'
-      from cleverhans.attacks_tf import jacobian_graph, jsma_batch
-
-      # Define Jacobian graph wrt to this input placeholder
-      preds = self.model.get_probs(x)
-      nb_classes = preds.get_shape().as_list()[-1]
-      grads = jacobian_graph(preds, x, nb_classes)
-
-      # Define appropriate graph (targeted / random target labels)
-      if self.y_target is not None:
-
-        def jsma_wrap(x_val, y_target):
-          return jsma_batch(
-              self.sess,
-              x,
-              preds,
-              grads,
-              x_val,
-              self.theta,
-              self.gamma,
-              self.clip_min,
-              self.clip_max,
-              nb_classes,
-              y_target=y_target)
-
-        # Attack is targeted, target placeholder will need to be fed
-        x_adv = tf.py_func(jsma_wrap, [x, self.y_target],
-                           self.tf_dtype)
-      else:
-
-        def jsma_wrap(x_val):
-          return jsma_batch(
-              self.sess,
-              x,
-              preds,
-              grads,
-              x_val,
-              self.theta,
-              self.gamma,
-              self.clip_min,
-              self.clip_max,
-              nb_classes,
-              y_target=None)
-
-        # Attack is untargeted, target values will be chosen at random
-        x_adv = tf.py_func(jsma_wrap, [x], self.tf_dtype)
-        x_adv.set_shape(x.get_shape())
-
-    return x_adv
-
-  def parse_params(self,
-                   theta=1.,
-                   gamma=1.,
-                   clip_min=0.,
-                   clip_max=1.,
-                   y_target=None,
-                   symbolic_impl=True,
-                   **kwargs):
-    """
-    Take in a dictionary of parameters and applies attack-specific checks
-    before saving them as attributes.
-
-    Attack-specific parameters:
-
-    :param theta: (optional float) Perturbation introduced to modified
-                  components (can be positive or negative)
-    :param gamma: (optional float) Maximum percentage of perturbed features
-    :param clip_min: (optional float) Minimum component value for clipping
-    :param clip_max: (optional float) Maximum component value for clipping
-    :param y_target: (optional) Target tensor if the attack is targeted
-    """
-    self.theta = theta
-    self.gamma = gamma
-    self.clip_min = clip_min
-    self.clip_max = clip_max
-    self.y_target = y_target
-    self.symbolic_impl = symbolic_impl
-
-    if len(kwargs.keys()) > 0:
-      warnings.warn("kwargs is unused and will be removed on or after "
-                    "2019-04-26.")
-
-    return True
-
-
-class VirtualAdversarialMethod(Attack):
-  """
-  This attack was originally proposed by Miyato et al. (2016) and was used
-  for virtual adversarial training.
-  Paper link: https://arxiv.org/abs/1507.00677
-
-  :param model: cleverhans.model.Model
-  :param sess: optional tf.Session
-  :param dtypestr: dtype of the data
-  :param kwargs: passed through to super constructor
-  """
-
-  def __init__(self, model, sess=None, dtypestr='float32', **kwargs):
-    """
-    Note: the model parameter should be an instance of the
-    cleverhans.model.Model abstraction provided by CleverHans.
-    """
-    if not isinstance(model, Model):
-      wrapper_warning_logits()
-      model = CallableModelWrapper(model, 'logits')
-
-    super(VirtualAdversarialMethod, self).__init__(model, sess, dtypestr,
-                                                   **kwargs)
-
-    self.feedable_kwargs = ('eps', 'xi', 'clip_min', 'clip_max')
-    self.structural_kwargs = ['num_iterations']
-
-  def generate(self, x, **kwargs):
-    """
-    Generate symbolic graph for adversarial examples and return.
-
-    :param x: The model's symbolic inputs.
-    :param kwargs: See `parse_params`
-    """
-    # Parse and save attack-specific parameters
-    assert self.parse_params(**kwargs)
-
-    return vatm(
-        self.model,
-        x,
-        self.model.get_logits(x),
-        eps=self.eps,
-        num_iterations=self.num_iterations,
-        xi=self.xi,
-        clip_min=self.clip_min,
-        clip_max=self.clip_max)
-
-  def parse_params(self,
-                   eps=2.0,
-                   nb_iter=None,
-                   xi=1e-6,
-                   clip_min=None,
-                   clip_max=None,
-                   num_iterations=None,
-                   **kwargs):
-    """
-    Take in a dictionary of parameters and applies attack-specific checks
-    before saving them as attributes.
-
-    Attack-specific parameters:
-
-    :param eps: (optional float )the epsilon (input variation parameter)
-    :param nb_iter: (optional) the number of iterations
-      Defaults to 1 if not specified
-    :param xi: (optional float) the finite difference parameter
-    :param clip_min: (optional float) Minimum input component value
-    :param clip_max: (optional float) Maximum input component value
-    :param num_iterations: Deprecated alias for `nb_iter`
-    """
-    # Save attack-specific parameters
-    self.eps = eps
-    if num_iterations is not None:
-      warnings.warn("`num_iterations` is deprecated. Switch to `nb_iter`."
-                    " The old name will be removed on or after 2019-04-26.")
-      # Note: when we remove the deprecated alias, we can put the default
-      # value of 1 for nb_iter back in the method signature
-      assert nb_iter is None
-      nb_iter = num_iterations
-    del num_iterations
-    if nb_iter is None:
-      nb_iter = 1
-    self.num_iterations = nb_iter
-    self.xi = xi
-    self.clip_min = clip_min
-    self.clip_max = clip_max
-    if len(kwargs.keys()) > 0:
-      warnings.warn("kwargs is unused and will be removed on or after "
-                    "2019-04-26.")
-    return True
 
 
 class CarliniWagnerL2(Attack):
@@ -881,46 +637,6 @@ class LBFGS_impl(object):
     return o_bestattack
 
 
-def vatm(model,
-         x,
-         logits,
-         eps,
-         back='tf',
-         num_iterations=1,
-         xi=1e-6,
-         clip_min=None,
-         clip_max=None):
-  """
-  A wrapper for the perturbation methods used for virtual adversarial
-  training : https://arxiv.org/abs/1507.00677
-  It calls the right function, depending on the
-  user's backend.
-
-  :param model: the model which returns the network unnormalized logits
-  :param x: the input placeholder
-  :param logits: the model's unnormalized output tensor
-  :param eps: the epsilon (input variation parameter)
-  :param num_iterations: the number of iterations
-  :param xi: the finite difference parameter
-  :param clip_min: optional parameter that can be used to set a minimum
-                  value for components of the example returned
-  :param clip_max: optional parameter that can be used to set a maximum
-                  value for components of the example returned
-  :return: a tensor for the adversarial example
-
-  """
-  assert back == 'tf'
-  # Compute VATM using TensorFlow
-  from cleverhans.attacks_tf import vatm as vatm_tf
-  return vatm_tf(
-      model,
-      x,
-      logits,
-      eps,
-      num_iterations=num_iterations,
-      xi=xi,
-      clip_min=clip_min,
-      clip_max=clip_max)
 
 
 class FastFeatureAdversaries(Attack):
